afternoon upload

+++ project/plato.discussionforum.webproject/WebContent/WEB-INF/jsp/approveThread.jsp
@@ -386,8 +386,8 @@
 												</c:forEach>
 												<td><c:out value="${thread.createdTime}" /></td>
 												<td><c:out value="${thread.modifiedTime}" /></td>
-												<td><a	href="saveApproveThread.html?thread_id=${thread.threadId}"><span class="fa fa-check" title="approve"></span></a>&nbsp;&nbsp;
-												<a	href="deleteThread.html?thread_id=${thread.threadId}&frmAprThr=1"><span class="fa fa-remove" title="delete"></span></a>
+												<td><a 	href="saveApproveThread.html?thread_id=${thread.threadId}"><span class="fa fa-check" title="approve"></span></a>&nbsp;&nbsp;
+												<a onclick="return confirm_delete(this);"	href="deleteThread.html?thread_id=${thread.threadId}&frmAprThr=1"><span class="fa fa-remove" title="delete"></span></a>
 												</td>
 											</tr>
 										</c:forEach>
@@ -600,6 +600,7 @@
     <script src="./resources/dist/js/app.min.js"></script>
     <!-- AdminLTE for demo purposes -->
     <script src="./resources/dist/js/demo.js"></script>
+    <script src="./resources/custom/js/custom.js"></script>
     <!-- page script -->
     <script>
       $(function () {
